Split operations.py up in to individual modules
