--- docs/api/operations.rst
@@ -1,6 +1,0 @@
-===================
-emsarray.operations
-===================
-
-.. automodule:: emsarray.operations
-    :members:
