+++ docs/api/operations/depth.rst
@@ -1,0 +1,6 @@
+=========================
+emsarray.operations.depth
+=========================
+
+.. automodule:: emsarray.operations.depth
+   :members:
