+++ docs/api/operations/index.rst
@@ -1,0 +1,23 @@
+===================
+emsarray.operations
+===================
+
+Operations are useful functions that
+transform datasets, extract subsets, manipulate geometry, or other useful actions.
+These functions do not depend on the specific format of the underlying dataset,
+and behave the same across all supported formats.
+
+.. toctree::
+   :hidden:
+   :glob:
+
+   ./*
+
+:mod:`emsarray.operations.depth`
+    These operations manipulate datasets with a depth axis,
+    such as the output of ocean models.
+
+:mod:`emsarray.operations.triangulate`
+    These operations triangulate the polygonal mesh of a dataset.
+    This is useful in combination with visualisation packages such as
+    Trimesh or Holoviews.
