+++ docs/api/operations/triangulate.rst
@@ -1,0 +1,6 @@
+===============================
+emsarray.operations.triangulate
+===============================
+
+.. automodule:: emsarray.operations.triangulate
+   :members:
